+++ tests/tests.py
@@ -1,0 +1,1 @@
+# TODO: Add tests for all the functions